fix(trusty-review)!: remove contributor profiling, hand it to tga
Deletes src/profile/ and cli_profile.rs, drops the default-on `profile`
Cargo feature and the `tga` + `rusqlite` dependencies it carried, and bumps
0.15.0 -> 0.16.0 (0.x breaking goes in the MINOR position).

Two things the issue did not name:

- pipeline/mapreduce/reduce.rs compiled unconditionally but imported
  crate::profile::synthesizer::jaccard_similarity, which only existed under
  the `profile` feature. `--no-default-features` was therefore already
  broken on main. The helper and both its tests move into reduce.rs beside
  their only caller.
- trusty-analyze's `trusty-review` dependency carried no
  `default-features = false`, so `--features review` resolved the whole
  default set. Removing `profile` would have broken a crate whose source
  nobody touched; that manifest fix ships here.

Closes #5466
Part of #5468
