+++ Cargo.toml
@@ -117,8 +117,11 @@ trusty-embedderd = { path = "crates/trusty-embedderd", version = "0.3.10" }
 # trusty-memory so `cargo install trusty-memory` produces both binaries from
 # one command (mirrors the trusty-embedderd / trusty-search pattern, PR #190).
 trusty-bm25-daemon = { path = "crates/trusty-bm25-daemon", version = "0.2.0" }
-# tga: trusty-git-analytics — developer productivity analytics.
-# Referenced by trusty-review for contributor-profile data pipeline (#558).
+# tga: trusty-git-analytics — developer productivity analytics, and (since
+# #5468) the sole owner of contributor profiling. No workspace member takes a
+# Cargo edge on it: #5466 removed trusty-review's, and the tga -> trusty-review
+# direction is a PATH lookup, not a manifest edge (#5236, DOC-67 §6). The entry
+# stays so a future in-tree consumer resolves from source rather than crates.io.
 tga = { path = "crates/trusty-git-analytics", version = "2.10" }
 # trusty-mpm: unified crate (replaces the former trusty-mpm-{core,mcp,client,
 # daemon,cli,tui,telegram} sub-crates). The gui sub-crate is kept separate
@@ -127,7 +130,7 @@ trusty-mpm = { path = "crates/trusty-mpm", version = "1.0.0" }
 # trusty-review: LLM-backed PR-review service. Referenced by trusty-analyze
 # (behind the optional `review` feature) so the analyze MCP server can expose
 # the trusty-review pipeline as `tr_review_*` tools (#630).
-trusty-review = { path = "crates/trusty-review", version = "0.15" }
+trusty-review = { path = "crates/trusty-review", version = "0.16" }
 # trusty-console: web dashboard for trusty services. As of #1318 the standalone
 # `trusty-console` crate is the SOLE producer of the `trusty-console` binary
 # (de-bundled from the host crates for single-owner-per-binary). It is still a
